+++ static/css/mystyle.css
@@ -102,6 +102,7 @@ li {
     top: 50%;
     transform: translateY(-50%); */
 	width: 10vh;
+	height: 6vh;
     border: none;
     background-color: #d7d7d7;
 	cursor: pointer;
